+++ gedcom/project03/validator.py
@@ -1,24 +1,38 @@
 import datetime
-import gedcom
+from model import Individual, Family
+import consts
 
 # TODO: Put all messages into one dict keyed by US##
 def get_message(id, item, args):
     pass
 
-def validate(object):
-    if isinstance(object, gedcom.Individual):
-        check_US07(object)
+def validate(obj):
+    if isinstance(obj, Individual):
+        check_US01(obj)
+        check_US07(obj)
     else:
-        pass
+        check_US01(obj)
 
-
+def check_US01(obj):
+    curr_date = datetime.datetime.now()
+    if isinstance(obj, Individual):
+        if obj.birth is not None and obj.birth > curr_date:
+            print(consts.MSG_US01.format("Birth", obj.birth))
+        elif obj.death is not None and obj.death > curr_date:
+            print(consts.MSG_US01.format("Death", obj.death))
+    elif isinstance(obj, Family):
+        if obj.marriage_date is not None and obj.marriage_date > curr_date:
+            print(consts.MSG_US01.format("Marriage", obj.marriage_date))
+        if obj.divorce_date is not None and obj.divorce_date > curr_date:
+            print(consts.MSG_US01.format("Divorce", obj.divorce_date))
 
 # Anomaly US07: Less then 150 years old
 def check_US07(individual) -> None:
     curr_date = datetime.datetime.now()
     valid = True
+
     if individual.birth is None:
-        print(f"Error US##: {str(individual)} has no birth date.") #TODO: This should be checked in a different ticket?
+        print("Error US##: {0} has no birth date.".format(str(individual))) #TODO: Should this be checked in a different issue?
         return
 
     if individual.death is None:
@@ -29,5 +43,5 @@ def check_US07(individual) -> None:
             valid = False
     
     if not valid:
-        print(f"Anomaly US07: {str(individual)} is more than 150 years old.")
+        print(consts.MSG_US07.format(str(individual)))
